+++ stroom-proxy/src/main/java/stroom/proxy/repo/ProxyRepositoryReader.java
@@ -180,14 +180,14 @@ public final class ProxyRepositoryReader {
                         continue;
                     }
 
-                    if (LOGGER.isInfoEnabled()) {
-                        LOGGER.info("run() - Cron Match at " + DateUtil.createNormalDateTimeString());
+                    if (LOGGER.isDebugEnabled()) {
+                        LOGGER.debug("Cron Match at " + DateUtil.createNormalDateTimeString());
                     }
 
                     try {
                         doRunWork();
                     } catch (final Throwable ex) {
-                        LOGGER.error("run() - Unhandled exception coming out of doRunWork()", ex);
+                        LOGGER.error("Unhandled exception coming out of doRunWork()", ex);
                     }
                 }
             }
@@ -195,7 +195,7 @@ public final class ProxyRepositoryReader {
             lock.unlock();
         }
 
-        LOGGER.info("run() - Completed ... Thread Exit");
+        LOGGER.info("Completed ... Thread Exit");
     }
 
     void doRunWork() {