+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#1137** : Change proxy remote url health check to accept a 406 code as the feed will not be specified.
+
 * Issue **#1135** : Data retention policies are now migrated to use `Type` and not `Stream Type`.