+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [v6.0-beta.39] - 2019-05-17
+
 * Issue **#1140** : Add health check for proxy feed status url.
 
 * Issue **#1138** : Stroom proxy now deletes empty repository directories based on creation time and depth first so that pruning empty directories is quicker and generally more successful.
@@ -1573,7 +1576,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v6.0-beta.38...6.0
+[Unreleased]: https://github.com/gchq/stroom/compare/v6.0-beta.39...6.0
+[v6.0-beta.39]: https://github.com/gchq/stroom/compare/v7.0-beta.38...v6.0-beta.39
 [v6.0-beta.38]: https://github.com/gchq/stroom/compare/v7.0-beta.36...v6.0-beta.38
 [v6.0-beta.37]: https://github.com/gchq/stroom/compare/v6.0-beta.36...v6.0-beta.37
 [v6.0-beta.36]: https://github.com/gchq/stroom/compare/v6.0-beta.35...v6.0-beta.36